mv resnet2dbb() from resnet.py to diverse_branch_block.py
